Updated snippets for lesson 4
