--- tutorials/nservicebus-step-by-step/4-publishing-events/Snippets/Core_8/Core_8.csproj
@@ -1,9 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <LangVersion>10.0</LangVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NServiceBus" Version="8.*" />
-  </ItemGroup>
-</Project>
