--- tutorials/nservicebus-step-by-step/4-publishing-events/Snippets/Core_8/Lesson4.cs
@@ -1,27 +1,0 @@
-﻿using System.Threading.Tasks;
-using NServiceBus;
-
-namespace Core_8
-{
-    #region Event
-    public class SomethingHappened :
-        IEvent
-    {
-        public string SomeProperty { get; set; }
-    }
-    #endregion
-
-    #region EventHandler
-    public class SomethingHappenedHandler :
-        IHandleMessages<SomethingHappened>
-    {
-        public Task Handle(SomethingHappened message, IMessageHandlerContext context)
-        {
-            // Do something with the event here
-
-            return Task.CompletedTask;
-        }
-    }
-    #endregion
-
-}
