+++ tutorials/nservicebus-step-by-step/4-publishing-events/Snippets/Snippets.sln
@@ -1,19 +1,24 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29728.190
+# Visual Studio Version 17
+VisualStudioVersion = 17.11.35303.130
 MinimumVisualStudioVersion = 15.0.26730.12
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core_8", "Core_8\Core_8.csproj", "{10CE989A-8E4F-400C-8BBD-3935B9FEE695}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core_9", "Core_9\Core_9.csproj", "{E6822F38-4323-4F72-89D7-9FBFA94FF5C2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{10CE989A-8E4F-400C-8BBD-3935B9FEE695}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{10CE989A-8E4F-400C-8BBD-3935B9FEE695}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E6822F38-4323-4F72-89D7-9FBFA94FF5C2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E6822F38-4323-4F72-89D7-9FBFA94FF5C2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E6822F38-4323-4F72-89D7-9FBFA94FF5C2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E6822F38-4323-4F72-89D7-9FBFA94FF5C2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B8EF01D2-E7BF-4004-8434-F859C9591272}
 	EndGlobalSection
 EndGlobal
